Minor Refactor: Separate NMA-Specific stuff from Nexus Web API Main Package
