+++ src/Games/NexusMods.Games.TestHarness/NexusMods.Games.TestHarness.csproj
@@ -8,6 +8,7 @@
         <ProjectReference Include="..\..\..\src\Networking\NexusMods.Networking.HttpDownloader\NexusMods.Networking.HttpDownloader.csproj" />
         <ProjectReference Include="..\..\..\src\Networking\NexusMods.Networking.NexusWebApi\NexusMods.Networking.NexusWebApi.csproj" />
         <ProjectReference Include="..\..\..\src\NexusMods.DataModel\NexusMods.DataModel.csproj" />
+        <ProjectReference Include="..\..\NexusMods.CLI\NexusMods.CLI.csproj" />
     </ItemGroup>
 
 </Project>
