+++ src/Networking/NexusMods.Networking.NexusWebApi.NMA/Types/NexusLoginJob.cs
@@ -1,8 +1,7 @@
-﻿using NexusMods.DataModel.Abstractions;
-using NexusMods.DataModel.Interprocess.Jobs;
+﻿using NexusMods.DataModel.Interprocess.Jobs;
 using NexusMods.DataModel.JsonConverters;
 
-namespace NexusMods.Networking.NexusWebApi.Types;
+namespace NexusMods.Networking.NexusWebApi.NMA.Types;
 
 [JsonName(nameof(NexusLoginJob))]
 public record NexusLoginJob : AJobEntity
