+++ src/Networking/NexusMods.Networking.NexusWebApi/Types/NXMUrl.cs
@@ -1,27 +1,6 @@
-using NexusMods.DataModel.Games;
 using System.Collections.Specialized;
 
 namespace NexusMods.Networking.NexusWebApi.Types;
-
-/// <summary>
-/// type of url
-/// </summary>
-// ReSharper disable once InconsistentNaming
-public enum NXMUrlType : byte
-{
-    /// <summary>
-    /// url referencing a mod file
-    /// </summary>
-    Mod = 0,
-    /// <summary>
-    /// url referencing a collection revision
-    /// </summary>
-    Collection,
-    /// <summary>
-    /// callback for oauth login
-    /// </summary>
-    OAuth
-}
 
 /// <summary>
 /// parsed url of nxm:// protocol, used in multiple ways for the site communicating with clients
@@ -165,144 +144,4 @@ public class NXMUrl
             return (NXMOAuthUrl)this;
         }
     }
-
-}
-
-/// <summary>
-/// specialized variant of NXMUrl for mod urls
-/// </summary>
-// ReSharper disable once InconsistentNaming
-public class NXMModUrl : NXMUrl
-{
-    /// <summary>
-    /// id of the mod page
-    /// </summary>
-    public ModId ModId { get; set; }
-    /// <summary>
-    /// id of the file (within that game domain)
-    /// </summary>
-    public FileId FileId { get; set; }
-    /// <summary>
-    /// game domain (name of the game within the Nexus Mods page)
-    /// </summary>
-    public GameDomain Game { get; set; }
-
-    /// <summary>
-    /// constructor
-    /// </summary>
-    /// <param name="uri">parsed uri</param>
-    /// <exception cref="ArgumentException">if the uri is not valid</exception>
-    public NXMModUrl(Uri uri)
-    {
-        UrlType = NXMUrlType.Mod;
-        if ((uri.Segments.Length != 5) || (uri.Segments[3] != "files/"))
-        {
-            throw new ArgumentException($"invalid nxm url \"{uri}\"");
-        }
-        Game = GameDomain.From(uri.Host);
-        try
-        {
-            ModId = ModId.From(ulong.Parse(uri.Segments[2].TrimEnd('/')));
-            FileId = FileId.From(ulong.Parse(uri.Segments[4].TrimEnd('/')));
-        }
-        catch (FormatException)
-        {
-            throw new ArgumentException($"invalid nxm url \"{uri}\"");
-        }
-    }
-
-    /// <summary>
-    /// serialize the url
-    /// </summary>
-    public override string ToString()
-    {
-        return $"nxm://{Game}/mods/{ModId}/files/{FileId}?{QueryString}";
-    }
-}
-
-/// <summary>
-/// specialized variant of NXMUrl for collecton links
-/// </summary>
-// ReSharper disable once InconsistentNaming
-public class NXMCollectionUrl : NXMUrl
-{
-    /// <summary>
-    /// string uniquely identifying a collection
-    /// </summary>
-    public CollectionSlug Slug { get; set; }
-    /// <summary>
-    /// the revision number (this is not the global revision id, the revision number is  like a version number within a collection)
-    /// </summary>
-    public RevisionNumber Revision { get; set; }
-    /// <summary>
-    /// name of the game within the Nexus Mods site
-    /// </summary>
-    public GameDomain Game { get; set; }
-
-    /// <summary>
-    /// constructor
-    /// </summary>
-    /// <param name="uri">parsed uri</param>
-    /// <exception cref="ArgumentException">if the uri is not valid</exception>
-    public NXMCollectionUrl(Uri uri)
-    {
-        UrlType = NXMUrlType.Collection;
-        if ((uri.Segments.Length != 5) || (uri.Segments[3] != "revisions/"))
-        {
-            throw new ArgumentException($"invalid nxm url \"{uri}\"");
-        }
-        Game = GameDomain.From(uri.Host);
-        try
-        {
-            Slug = CollectionSlug.From(uri.Segments[2].TrimEnd('/'));
-            Revision = RevisionNumber.From(ulong.Parse(uri.Segments[4].TrimEnd('/')));
-        }
-        catch (FormatException)
-        {
-            throw new ArgumentException($"invalid nxm url \"{uri}\"");
-        }
-    }
-
-    /// <summary>
-    /// serialize the url
-    /// </summary>
-    public override string ToString()
-    {
-        return $"nxm://{Game}/collections/{Slug}/revisions/{Revision}?{QueryString}";
-    }
-}
-
-/// <summary>
-/// specialized variant of NXMUrl for oauth callback links
-/// </summary>
-// ReSharper disable once IdentifierTypo
-// ReSharper disable once InconsistentNaming
-public class NXMOAuthUrl : NXMUrl
-{
-    /// <summary>
-    /// a code. This gets used in another request to get at the actual authorization code
-    /// </summary>
-    public string? Code => Query.Get("code");
-
-    /// <summary>
-    /// this is a random id we generated when initially creating the request, it allows us to
-    /// match the respons to the initial request we made
-    /// </summary>
-    public string? State => Query.Get("state");
-
-    /// <summary>
-    /// constructor
-    /// </summary>
-    /// <param name="uri">parsed url</param>
-    public NXMOAuthUrl(Uri uri)
-    {
-        UrlType = NXMUrlType.OAuth;
-        if ((uri.Segments.Length != 2) || (uri.Segments[1] != "callback"))
-            throw new ArgumentException($"invalid nxm url \"{uri}\"");
-    }
-
-    /// <summary>
-    /// serialize the url
-    /// </summary>
-    public override string ToString() => $"nxm://oauth/callback?{QueryString}";
 }
