+++ src/Networking/NexusMods.Networking.NexusWebApi/Types/NXMUrlType.cs
@@ -1,0 +1,21 @@
+namespace NexusMods.Networking.NexusWebApi.Types;
+
+/// <summary>
+/// type of url
+/// </summary>
+// ReSharper disable once InconsistentNaming
+public enum NXMUrlType : byte
+{
+    /// <summary>
+    /// url referencing a mod file
+    /// </summary>
+    Mod = 0,
+    /// <summary>
+    /// url referencing a collection revision
+    /// </summary>
+    Collection,
+    /// <summary>
+    /// callback for oauth login
+    /// </summary>
+    OAuth
+}
