+++ src/NexusMods.App.UI/LeftMenu/Items/LaunchButtonViewModel.cs
@@ -2,18 +2,11 @@ using System.Collections.ObjectModel;
 using System.Reactive;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
-using System.Windows.Input;
 using DynamicData;
-using DynamicData.Aggregation;
-using DynamicData.PLinq;
-using NexusMods.App.UI.Extensions;
-using NexusMods.CLI.Verbs;
-using NexusMods.DataModel.Abstractions;
 using NexusMods.DataModel.Games;
 using NexusMods.DataModel.Interprocess.Jobs;
 using NexusMods.DataModel.Interprocess.Messages;
 using NexusMods.DataModel.Loadouts;
-using NexusMods.DataModel.Loadouts.Cursors;
 using NexusMods.DataModel.Loadouts.Markers;
 using NexusMods.DataModel.RateLimiting;
 using ReactiveUI;
