+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -96,6 +96,7 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\Networking\NexusMods.Networking.NexusWebApi.NMA\NexusMods.Networking.NexusWebApi.NMA.csproj" />
         <ProjectReference Include="..\Networking\NexusMods.Networking.NexusWebApi\NexusMods.Networking.NexusWebApi.csproj" />
         <ProjectReference Include="..\NexusMods.DataModel\NexusMods.DataModel.csproj" />
     </ItemGroup>
