--- src/NexusMods.CLI/Types/IProtocolHandler.cs
@@ -1,7 +1,0 @@
-namespace NexusMods.CLI.Types;
-
-public interface IProtocolHandler
-{
-    public string Protocol { get; }
-    public Task Handle(string url, CancellationToken token);
-}
