+++ src/NexusMods.CLI/Types/IpcHandlers/NxmIpcProtocolHandler.cs
@@ -1,14 +1,15 @@
-using NexusMods.CLI.Types;
 using NexusMods.DataModel.Interprocess;
+using NexusMods.Networking.NexusWebApi;
+using NexusMods.Networking.NexusWebApi.NMA.Messages;
 using NexusMods.Networking.NexusWebApi.Types;
 
-namespace NexusMods.Networking.NexusWebApi;
+namespace NexusMods.CLI.Types.IpcHandlers;
 
 /// <summary>
 /// a handler for nxm:// urls
 /// </summary>
 // ReSharper disable once InconsistentNaming
-public class NXMProtocolHandler : IProtocolHandler
+public class NxmIpcProtocolHandler : IIpcProtocolHandler
 {
     /// <inheritdoc/>
     public string Protocol { get; } = "nxm";
@@ -18,7 +19,7 @@ public class NXMProtocolHandler : IProtocolHandler
     /// <summary>
     /// constructor
     /// </summary>
-    public NXMProtocolHandler(IMessageProducer<NXMUrlMessage> messages)
+    public NxmIpcProtocolHandler(IMessageProducer<NXMUrlMessage> messages)
     {
         _messages = messages;
     }
