+++ tests/Games/NexusMods.Games.TestFramework/TestModDownloader.cs
@@ -4,6 +4,7 @@ using NexusMods.DataModel.Games;
 using NexusMods.Games.TestFramework.Downloader;
 using NexusMods.Networking.HttpDownloader;
 using NexusMods.Networking.NexusWebApi;
+using NexusMods.Networking.NexusWebApi.NMA.Extensions;
 using NexusMods.Paths;
 
 namespace NexusMods.Games.TestFramework;
@@ -88,7 +89,7 @@ public class TestModDownloader
     {
         // TODO: Nexus Web API needs To Support IFileSystem at some point.
         _logger.LogInformation("Downloading {ModId} {FileId} {Hash}", nexusMod.ModId, nexusMod.FileId, nexusMod.Hash);
-        var uris = await _nexusClient.DownloadLinks(GameDomain.Cyberpunk2077, nexusMod.ModId, nexusMod.FileId, token);
+        var uris = await _nexusClient.DownloadLinksAsync(GameDomain.Cyberpunk2077, nexusMod.ModId, nexusMod.FileId, token);
         var downloadUris = uris.Data.Select(u => new HttpRequestMessage(HttpMethod.Get, u.Uri)).ToArray();
         var downloadHash = await _httpDownloader.DownloadAsync(downloadUris, folderPath, token: token);
         if (downloadHash != nexusMod.Hash)
