+++ tests/Networking/NexusMods.Networking.NexusWebApi.Tests/OAuth2MessageFactoryTests.cs
@@ -9,6 +9,9 @@ using Moq.Protected;
 using System.Text.Json;
 using NexusMods.Common.OSInterop;
 using NexusMods.DataModel.Interprocess.Jobs;
+using NexusMods.Networking.NexusWebApi.DTOs;
+using NexusMods.Networking.NexusWebApi.NMA;
+using NexusMods.Networking.NexusWebApi.NMA.Messages;
 
 namespace NexusMods.Networking.NexusWebApi.Tests;
