+++ tests/Networking/NexusMods.Networking.NexusWebApi.Tests/OAuthTests.cs
@@ -9,6 +9,9 @@ using System.Text.Json;
 using FluentAssertions;
 using NexusMods.Common.OSInterop;
 using NexusMods.DataModel.Interprocess.Jobs;
+using NexusMods.Networking.NexusWebApi.DTOs;
+using NexusMods.Networking.NexusWebApi.NMA;
+using NexusMods.Networking.NexusWebApi.NMA.Messages;
 
 namespace NexusMods.Networking.NexusWebApi.Tests;
